move LightningAdapter to pytorch.lightning package
