+++ examples/computer_vision/mnist_pl/model_def.py
@@ -4,7 +4,8 @@ interface to build a basic MNIST network. LightningAdapter utilizes the provided
 LightningModule with Determined's PyTorch control loop.
 """
 
-from determined.pytorch import PyTorchTrialContext, DataLoader, LightningAdapter
+from determined.pytorch import PyTorchTrialContext, DataLoader
+from determined.pytorch.lightning import LightningAdapter
 import mnist
 
 class MNISTTrial(LightningAdapter):
