--- harness/determined/pytorch/_lightning/__init__.py
@@ -1,1 +1,0 @@
-from determined.pytorch._lightning._adapter import LightningAdapter
