+++ harness/determined/pytorch/lightning/__init__.py
@@ -1,0 +1,1 @@
+from determined.pytorch.lightning._adapter import LightningAdapter
