+++ harness/determined/pytorch/lightning/_adapter.py
